move objective function to a seperate file
